+++ lib/postgres.dart
@@ -1,4 +1,3 @@
 export 'package:postgres/postgres.dart' show SslMode;
 
 export 'src/postgres/postgres_driver.dart' show PostgresDriver;
-export 'src/postgres/returning.dart' show applyReturningClause;