+++ lib/src/dialect/rewritten_sql.dart
@@ -1,7 +1,0 @@
-class RewrittenSql {
-  final String sql;
-
-  final Object? parameters;
-
-  const RewrittenSql(this.sql, this.parameters);
-}