+++ lib/src/migration.dart
@@ -1,2 +1,0 @@
-export 'migration/migration.dart';
-export 'migration/migrator.dart';