+++ lib/src/postgres/postgres_session.dart
@@ -2,7 +2,7 @@ import 'package:postgres/postgres.dart';
 
 import '../driver/query_result.dart';
 import '../driver/ratel_session.dart';
-import 'query_result_mapper.dart';
+import 'postgres_result_mapper.dart';
 
 class PostgresSession implements RatelSession {
   final TxSession _tx;
@@ -15,6 +15,6 @@ class PostgresSession implements RatelSession {
     final result = (parameters == null || parameters.isEmpty)
         ? await _tx.execute(sql)
         : await _tx.execute(Sql.named(sql), parameters: parameters);
-    return toQueryResult(result);
+    return PostgresResultMapper.toQueryResult(result);
   }
 }