+++ lib/src/postgres/query_result_mapper.dart
@@ -1,8 +1,0 @@
-import 'package:postgres/postgres.dart';
-
-import '../driver/query_result.dart';
-
-QueryResult toQueryResult(Result result) => QueryResult(
-      rows: [for (final row in result) row.toColumnMap()],
-      affectedRows: result.affectedRows,
-    );