+++ lib/src/query.dart
@@ -1,2 +1,0 @@
-export 'query/built_query.dart';
-export 'query/query.dart';